+++ src/Panama/ViewModel/Title/TitleQueueController.cs
@@ -1,4 +1,5 @@
-﻿using Restless.Panama.Database.Tables;
+﻿using Restless.Panama.Core;
+using Restless.Panama.Database.Tables;
 using Restless.Panama.Resources;
 using Restless.Toolkit.Controls;
 using System.Data;
